Fix GetPieceAtSquare argument type

+++ move.go
@@ -39,16 +39,16 @@ func (m *Move) Validate(board *Board) bool {
 	m.isCapture = false
 	m.isPromotion = false
 
-	piece := board.GetPieceAtSquare(uint8(m.From))
+	piece := board.GetPieceAtSquare(m.From)
 	if piece == WHITE_PAWN {
 		if m.To == m.From+8 {
 			m.validated = true
 		} else if m.To == m.From+16 && m.From >= A2 && m.From <= H2 {
 			m.validated = true
-		} else if m.To == m.From+7 && m.From%8 != 0 && board.GetPieceAtSquare(uint8(m.To)) != EMPTY {
+		} else if m.To == m.From+7 && m.From%8 != 0 && board.GetPieceAtSquare(m.To) != EMPTY {
 			m.validated = true
 			m.isCapture = true
-		} else if m.To == m.From+9 && m.From%8 != 7 && board.GetPieceAtSquare(uint8(m.To)) != EMPTY {
+		} else if m.To == m.From+9 && m.From%8 != 7 && board.GetPieceAtSquare(m.To) != EMPTY {
 			m.validated = true
 			m.isCapture = true
 		}
